requirements.txt
# LangChain 核心
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.10
langchain-core>=0.1.0

# 向量数据库
chromadb>=0.4.0
faiss-cpu>=1.7.4

# 文档处理
pypdf>=3.17.0
python-docx>=0.8.11
beautifulsoup4>=4.12.0
markdown>=3.5.0

# 网络请求
requests>=2.31.0
aiohttp>=3.9.0
httpx>=0.25.0

# 数据处理
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0

# 配置管理
pydantic>=2.5.0
python-dotenv>=1.0.0

# 日志和监控
loguru>=0.7.0

# 异步支持
asyncio-mqtt>=0.16.0
redis>=5.0.0

# 工具集成
openai>=1.3.0
tiktoken>=0.5.0

# 开发工具
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0

# 其他工具
tqdm>=4.66.0
click>=8.1.0
rich>=13.0.0
psycopg2-binary>=2.9.0
alembic>=1.7.0

# 工具模块依赖
PyYAML>=6.0
scikit-learn>=1.3.0
sentence-transformers>=2.2.0

# 代码执行和数据分析
ipython>=8.0.0
duckdb>=0.9.0
plotly>=5.15.0
dash>=2.14.0
scipy>=1.10.0

# 异步和并发
nest-asyncio>=1.5.0

# 其他工具依赖
PyYAML>=6.0

# 代码执行器依赖
ipykernel>=6.0.0
jupyter>=1.0.0